+++ client/context/query.go
@@ -83,10 +83,6 @@ func (ctx CLIContext) query(path string, key cmn.HexBytes) (res []byte, height i
 		return res, height, err
 	}
 
-	if ctx.Height <= 1 && !ctx.TrustNode {
-		return res, height, errors.New("cannot query with proof when height <= 1; please provide a valid height")
-	}
-
 	opts := rpcclient.ABCIQueryOptions{
 		Height: ctx.Height,
 		Prove:  !ctx.TrustNode,
@@ -99,7 +95,7 @@ func (ctx CLIContext) query(path string, key cmn.HexBytes) (res []byte, height i
 
 	resp := result.Response
 	if !resp.IsOK() {
-		return res, height, errors.New(resp.Log)
+		return res, resp.Height, errors.New(resp.Log)
 	}
 
 	// data from trusted node or subspace query doesn't need verification
@@ -109,7 +105,7 @@ func (ctx CLIContext) query(path string, key cmn.HexBytes) (res []byte, height i
 
 	err = ctx.verifyProof(path, resp)
 	if err != nil {
-		return res, height, err
+		return res, resp.Height, err
 	}
 
 	return resp.Value, resp.Height, nil