+++ server/start.go
@@ -100,7 +100,6 @@ func startStandAlone(ctx *Context, appCreator AppCreator) error {
 	// run forever (the node will not be returned)
 	select {}
 
-	return nil
 }
 
 func startInProcess(ctx *Context, appCreator AppCreator) (*node.Node, error) {